perform_upload: Add option to not force revision to 'head' if not specified
This commit introduces a new parameter named 'create_additional_revision'
allowing new revision to be created when uploading data in an item that
already has a revision.

Ideally the revision should never be forced to 'head', the parameter
'create_additional_revision' has been introduced to maintain backward
compatibility while providing a way for client to get the expected
behavior.